feat(core): add WatchEvent struct
