+++ packages/nx/src/native/tests/native.spec.ts
@@ -37,4 +37,4 @@ describe('watcher', () => {
   }, 10_000_000);
 });
 
-///..
+///...
