+++ packages/nx/src/native/watcher/get_ignore_files.rs
@@ -1,0 +1,27 @@
+use ignore::WalkBuilder;
+use ignore_files::IgnoreFile;
+use std::path::PathBuf;
+
+pub fn get_ignore_files<T: AsRef<str>>(root: T) -> Vec<IgnoreFile> {
+    let root = root.as_ref();
+
+    let mut walker = WalkBuilder::new(root);
+    walker.hidden(false);
+
+    walker
+        .build()
+        .flatten()
+        .filter(|result| {
+            result.path().ends_with(".nxignore") || result.path().ends_with(".gitignore")
+        })
+        .map(|result| {
+            let path: PathBuf = result.path().into();
+            let parent: PathBuf = path.parent().unwrap_or(&path).into();
+            IgnoreFile {
+                path,
+                applies_in: Some(parent),
+                applies_to: None,
+            }
+        })
+        .collect()
+}
