+++ packages/nx/src/native/watcher/watch_filterer.rs
@@ -1,0 +1,51 @@
+use watchexec::error::RuntimeError;
+use watchexec::filter::Filterer;
+use watchexec_events::filekind::{FileEventKind, ModifyKind};
+use watchexec_events::{Event, FileType, Priority, Source, Tag};
+use watchexec_filterer_ignore::IgnoreFilterer;
+
+#[derive(Debug)]
+pub struct WatchFilterer {
+    pub inner: IgnoreFilterer,
+}
+
+/// Used to filter out events that that come from watchexec
+impl Filterer for WatchFilterer {
+    fn check_event(&self, event: &Event, priority: Priority) -> Result<bool, RuntimeError> {
+        //
+        // Tags will be a Vec that contains multiple types of information for a given event
+        // We are only interested if:
+        // 1) A `FileEventKind` is modified, created, removed, or renamed
+        // 2) A Path that is a FileType::File
+        // 3) Deleted files do not have a FileType::File (because they're deleted..), check if a path is valid
+        // 4) Only FileSystem sources are valid
+        // If there's a tag that doesnt confine to this criteria, we `return` early, otherwise we `continue`.
+        for tag in &event.tags {
+            match tag {
+                Tag::FileEventKind(file_event) => match file_event {
+                    FileEventKind::Modify(ModifyKind::Name(_)) => continue,
+                    FileEventKind::Modify(ModifyKind::Data(_)) => continue,
+                    FileEventKind::Create(_) => continue,
+                    FileEventKind::Remove(_) => continue,
+                    _ => return Ok(false),
+                },
+                Tag::Path {
+                    file_type: Some(FileType::File),
+                    ..
+                } => continue,
+                Tag::Path {
+                    path,
+                    file_type: _file_type,
+                } if !path.display().to_string().ends_with('~') => continue,
+                Tag::Source(Source::Filesystem) => continue,
+                _ => return Ok(false),
+            }
+        }
+
+        if !self.inner.check_event(event, priority)? {
+            return Ok(false);
+        }
+
+        Ok(true)
+    }
+}
